app runs successfully with new changes to auth endpoints but hasn't been manually tested or fully linked to frontend yet
