+++ api/src/auth/auth_funcs.py
@@ -1,13 +1,8 @@
-import datetime
-import json
 import sqlite3
 
-from flask import request, make_response, jsonify, current_app as app
-from functools import wraps
+from flask import request
 from werkzeug.security import generate_password_hash, check_password_hash
-from datetime import datetime, timedelta, timezone
-from flask_jwt_extended import create_access_token, get_jwt, get_jwt_identity, \
-                               unset_jwt_cookies, jwt_required, JWTManager
+from flask_jwt_extended import create_access_token
 
 from db_connection import connect_to_db
 
@@ -30,35 +25,7 @@ def signup_user():
         return message
 
     return message
-
-
-@app.after_request
-def refresh_expiring_jwt(response):
-    """
-    Using an `after_request` callback:
-    Refresh any token that is within 30 minutes of expiring.
-    Takes the response of the protected endpoint as an argument.
-    """
-    try:
-        exp_timestamp = get_jwt()["exp"]
-        now = datetime.now(timezone.utc)
-        target_timestamp = datetime.timestamp(now + timedelta(minutes=30))
-        if target_timestamp > exp_timestamp:
-            access_token = create_access_token(identity=get_jwt_identity())
-            data = response.get_json()
-            if type(data) is dict:
-                data["access_token"] = access_token
-                response.data = json.dumps(data)
-                # or try to use below instead of above line...
-                # set_access_cookies(response, access_token)
-        return response
-    except (RuntimeError, KeyError):
-        """
-        Case where there is not a valid JWT. 
-        Just return the original response.
-        """
-        return response
-
+    
 
 def login_user():
     try:
